Python/Virtual-Assistant/requirements.txt
gunicorn
Pillow
requests
beautifulsoup4
pyttsx3
spacy
google-generativeai
protobuf
SpeechRecognition
# resend
python-dotenv
Flask
flask-cors
apscheduler
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
